change to domain-driven project structure
