+++ application-contracts/src/main/java/com/pharmacy/application/contracts/dtos/CountManuProdsDTO.java
@@ -1,0 +1,4 @@
+package com.pharmacy.application.contracts.dtos;
+
+public record CountManuProdsDTO(String name, long count) {
+}
