+++ application-contracts/src/main/java/com/pharmacy/application/contracts/dtos/ProductDTO.java
@@ -1,0 +1,7 @@
+package com.pharmacy.application.contracts.dtos;
+
+import java.util.Date;
+import java.util.UUID;
+
+public record ProductDTO(UUID id, String name, ManufacturerDTO manufacturer, int price, Date expiryDate) {
+}
