+++ application/src/main/java/com/pharmacy/application/exceptions/errors/ProductExistsException.java
@@ -1,0 +1,7 @@
+package com.pharmacy.application.exceptions.errors;
+
+public class ProductExistsException extends RuntimeException {
+    public ProductExistsException(String message) {
+        super(message);
+    }
+}
