+++ application/src/main/java/com/pharmacy/application/exceptions/errors/ProductNotFoundException.java
@@ -1,0 +1,7 @@
+package com.pharmacy.application.exceptions.errors;
+
+public class ProductNotFoundException extends RuntimeException {
+    public ProductNotFoundException(String message) {
+        super(message);
+    }
+}
