+++ application/src/main/java/com/pharmacy/application/exceptions/errors/WrongManufacturerException.java
@@ -1,0 +1,7 @@
+package com.pharmacy.application.exceptions.errors;
+
+public class WrongManufacturerException extends RuntimeException {
+    public WrongManufacturerException(String message) {
+        super(message);
+    }
+}
