--- src/main/java/com/example/pharmacyback/PharmacyBackApplication.java
@@ -1,12 +1,0 @@
-package com.example.pharmacyback;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class PharmacyBackApplication {
-
-  public static void main(String[] args) {
-    SpringApplication.run(PharmacyBackApplication.class, args);
-  }
-}
