--- src/main/java/com/example/pharmacyback/dto/CountManuProdDTO.java
@@ -1,3 +1,0 @@
-package com.example.pharmacyback.dto;
-
-public record CountManuProdDTO(String name, long count) {}
