--- src/main/java/com/example/pharmacyback/exceptions/ExceptionResponse.java
@@ -1,5 +1,0 @@
-package com.example.pharmacyback.exceptions;
-
-import org.springframework.http.HttpStatus;
-
-public record ExceptionResponse(String message, HttpStatus status) {}
