--- src/main/java/com/example/pharmacyback/exceptions/errors/DateExpiredException.java
@@ -1,7 +1,0 @@
-package com.example.pharmacyback.exceptions.errors;
-
-public class DateExpiredException extends RuntimeException {
-  public DateExpiredException(String message) {
-    super(message);
-  }
-}
