--- src/main/java/com/example/pharmacyback/exceptions/errors/ProductExistsException.java
@@ -1,7 +1,0 @@
-package com.example.pharmacyback.exceptions.errors;
-
-public class ProductExistsException extends RuntimeException {
-  public ProductExistsException(String message) {
-    super(message);
-  }
-}
