--- src/main/java/com/example/pharmacyback/exceptions/errors/ProductNotFoundException.java
@@ -1,7 +1,0 @@
-package com.example.pharmacyback.exceptions.errors;
-
-public class ProductNotFoundException extends RuntimeException {
-  public ProductNotFoundException(String message) {
-    super(message);
-  }
-}
