--- src/main/java/com/example/pharmacyback/exceptions/errors/WrongManudacturerException.java
@@ -1,7 +1,0 @@
-package com.example.pharmacyback.exceptions.errors;
-
-public class WrongManudacturerException extends RuntimeException {
-  public WrongManudacturerException(String message) {
-    super(message);
-  }
-}
