--- src/test/java/com/example/pharmacyback/PharmacyBackApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.pharmacyback;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class PharmacyBackApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
